Offline & different user schedule support.
All loaded schedules will be stored, they will be displayed when no
connection can be made to the server in time. It also features an
offline search engine to search through the offline schedules.

The goal was to not make the user feel that his internet connection
sucks. I think I did a pretty good job at that.